25/11/sua be theo yeu cau tu sep, import du lieu mau

+++ back-end/controller/BookController.js
@@ -1,30 +1,69 @@
 const bookService = require("../service/BookService");
-
+// Hàm formatBooks định dạng sách
+// dùng cho các hàm : getAll,getByCategory, getByAuthor, getByTitle, getAllSortedByPrice,
+//  searchBooks, getLowStockBooks
+const formatBooks = (books) => {
+  return books.map((book) => ({
+    id: book._id, 
+    title: book.title,
+    author: book.author,
+    category: book.categoryID ? {
+        id: book.categoryID._id, // ID danh mục
+        name: book.categoryID.name // Tên danh mục
+    } : null, // Nếu không có categoryID, trả về null
+    publisher: book.publisherID ? {
+        id: book.publisherID._id, // ID nhà xuất bản
+        name: book.publisherID.name // Tên nhà xuất bản
+    } : null, // Nếu không có publisherID, trả về null
+    description: book.description,
+    price: book.price,
+    stock: book.stock,
+    sale: book.sales,
+    view: book.views,
+    cover_image: book.cover_image,
+    created_at: book.created_at,   // Thêm trường created_at
+    updated_at: book.updated_at
+  }));
+};
+// Hàm formatBooks định dạng sách dùng cho hàm: getHotProducts, getFeaturedProducts, getLatestBooks
+const formatProducts = (products) => {          
+  return products.map((product) => ({
+    id: product._id, 
+    title: product.title,
+    author: product.author,
+    category: product.categoryID ? {
+        id: product.categoryID._id, // ID danh mục
+        name: product.categoryID.name // Tên danh mục
+    } : null, // Nếu không có categoryID, trả về null
+    publisher: product.publisherID ? {
+        id: product.publisherID._id, // ID nhà xuất bản
+        name: product.publisherID.name // Tên nhà xuất bản
+    } : null, // Nếu không có publisherID, trả về null
+    description: product.description,
+    price: product.price,
+    stock: product.stock,
+    view: product.views,
+    cover_image: product.cover_image,
+    created_at: product.created_at,   // Thêm trường created_at
+    updated_at: product.updated_at
+  }));
+};
 exports.getAll = async () => {
   try {
     const books = await bookService.getAll();
-
     // Định dạng dữ liệu trước khi trả về
-    const formattedBooks = books.map((book) => ({
-      id: book._id, 
-      title: book.title,
-      author: book.author,
-      category: book.categoryID ? {
-          id: book.categoryID._id, // ID danh mục
-          name: book.categoryID.name // Tên danh mục
-      } : null, // Nếu không có categoryID, trả về null
-      publisher: book.publisherID ? {
-          id: book.publisherID._id, // ID nhà xuất bản
-          name: book.publisherID.name // Tên nhà xuất bản
-      } : null, // Nếu không có publisherID, trả về null
-      description: book.description,
-      price: book.price,
-      stock: book.stock,
-      cover_image: book.cover_image,
-      created_at: book.created_at,   // Thêm trường created_at
-      updated_at: book.updated_at
-    }));
-
+    const formattedBooks = formatBooks(books);
+    // Trả về danh sách sách đã được định dạng
+    return formattedBooks; // Đảm bảo trả về danh sách đã định dạng
+  } catch (error) {
+    throw new Error('Error fetching books: ' + error.message);
+  }
+};
+exports.getAllvisible = async () => {
+  try {
+    const books = await bookService.getAllvisible();
+    // Định dạng dữ liệu trước khi trả về
+    const formattedBooks = formatBooks(books);
     // Trả về danh sách sách đã được định dạng
     return formattedBooks; // Đảm bảo trả về danh sách đã định dạng
   } catch (error) {
@@ -40,25 +79,7 @@ exports.getByCategory = async (categoryID) => {
     // Gọi hàm từ service để lấy sách theo thể loại
     const books = await bookService.getByCategoryID(categoryID);
      // Định dạng dữ liệu trước khi trả về
-     const formattedBooks = books.map((book) => ({
-      id: book._id, // ObjectId của sách
-      title: book.title,
-      author: book.author,
-      category: book.categoryID ? {
-          id: book.categoryID._id, // ID danh mục
-          name: book.categoryID.name // Tên danh mục
-      } : null, // Nếu không có categoryID, trả về null
-      publisher: book.publisherID ? {
-          id: book.publisherID._id, // ID nhà xuất bản
-          name: book.publisherID.name // Tên nhà xuất bản
-      } : null, // Nếu không có publisherID, trả về null
-      description: book.description,
-      price: book.price,
-      stock: book.stock,
-      cover_image: book.cover_image,
-      created_at: book.created_at,   // Thêm trường created_at
-      updated_at: book.updated_at
-    }));
+     const formattedBooks = formatBooks(books);
 
     // Trả về danh sách sách đã được định dạng
     return formattedBooks; // Đảm bảo trả về danh sách đã định dạng
@@ -66,91 +87,14 @@ exports.getByCategory = async (categoryID) => {
     throw new Error('Error fetching books by category: ' + error.message);
   }
 };
-exports.getLowStockBooks = async () => {
 
-  try {
-
-    const books = await bookService.getLowStockBooks();
-
-    
-
-    // Định dạng lại dữ liệu trước khi trả về
-
-    const formattedBooks = books.map((book) => ({
-
-      id: book._id, // ObjectId của sách
-
-      title: book.title,
-
-      author: book.author,
-
-      category: book.categoryID ? {
-
-          id: book.categoryID._id, // ID danh mục
-
-          name: book.categoryID.name // Tên danh mục
-
-      } : null, // Nếu không có categoryID, trả về null
-
-      publisher: book.publisherID ? {
-
-          id: book.publisherID._id, // ID nhà xuất bản
-
-          name: book.publisherID.name // Tên nhà xuất bản
-
-      } : null, // Nếu không có publisherID, trả về null
-
-      description: book.description,
-
-      price: book.price,
-
-      stock: book.stock,
-
-      cover_image: book.cover_image,
-
-      created_at: book.created_at,   // Thêm trường created_at
-
-      updated_at: book.updated_at
-
-    }));
-
-
-
-    return formattedBooks;
-
-  } catch (error) {
-
-    throw new Error('Error fetching low stock books: ' + error.message);
-
-  }
-
-};
 // Hàm lấy sách theo tác giả
 exports.getByAuthor = async (author) => {
   try {
     // Gọi hàm từ service để lấy sách theo tác giả
     const books = await bookService.getByAuthor(author);
     // Định dạng dữ liệu trước khi trả về
-    const formattedBooks = books.map((book) => ({
-      id: book._id, // ObjectId của sách
-      title: book.title,
-      author: book.author,
-      category: book.categoryID ? {
-          id: book.categoryID._id, // ID danh mục
-          name: book.categoryID.name // Tên danh mục
-      } : null, // Nếu không có categoryID, trả về null
-      publisher: book.publisherID ? {
-          id: book.publisherID._id, // ID nhà xuất bản
-          name: book.publisherID.name // Tên nhà xuất bản
-      } : null, // Nếu không có publisherID, trả về null
-      description: book.description,
-      price: book.price,
-      stock: book.stock,
-      cover_image: book.cover_image,
-      created_at: book.created_at,   // Thêm trường created_at
-      updated_at: book.updated_at
-    }));
-
+    const formattedBooks = formatBooks(books);
     // Trả về danh sách sách đã được định dạng
     return formattedBooks; // Đảm bảo trả về danh sách đã định dạng
   } catch (error) {
@@ -163,26 +107,7 @@ exports.getByTitle = async (title) => {
     // Gọi hàm từ service để lấy sách theo tên từ database
     const books = await bookService.getByTitle(title);
   // Định dạng dữ liệu trước khi trả về
-  const formattedBooks = books.map((book) => ({
-    id: book._id, // ObjectId của sách
-    title: book.title,
-    author: book.author,
-    category: book.categoryID ? {
-        id: book.categoryID._id, // ID danh mục
-        name: book.categoryID.name // Tên danh mục
-    } : null, // Nếu không có categoryID, trả về null
-    publisher: book.publisherID ? {
-        id: book.publisherID._id, // ID nhà xuất bản
-        name: book.publisherID.name // Tên nhà xuất bản
-    } : null, // Nếu không có publisherID, trả về null
-    description: book.description,
-    price: book.price,
-    stock: book.stock,
-    cover_image: book.cover_image,
-    created_at: book.created_at,   // Thêm trường created_at
-    updated_at: book.updated_at
-  }));
-
+  const formattedBooks = formatBooks(books);
   // Trả về danh sách sách đã được định dạng
   return formattedBooks; // Đảm bảo trả về danh sách đã định dạng
   } catch (error) {
@@ -196,26 +121,7 @@ exports.getAllSortedByPrice = async (sortOrder) => {
     // Gọi hàm từ service để lấy sách và sắp xếp theo giá
     const books = await bookService.getAllSortedByPrice(sortOrder);
      // Định dạng dữ liệu trước khi trả về
-     const formattedBooks = books.map((book) => ({
-      id: book._id, // ObjectId của sách
-      title: book.title,
-      author: book.author,
-      category: book.categoryID ? {
-          id: book.categoryID._id, // ID danh mục
-          name: book.categoryID.name // Tên danh mục
-      } : null, // Nếu không có categoryID, trả về null
-      publisher: book.publisherID ? {
-          id: book.publisherID._id, // ID nhà xuất bản
-          name: book.publisherID.name // Tên nhà xuất bản
-      } : null, // Nếu không có publisherID, trả về null
-      description: book.description,
-      price: book.price,
-      stock: book.stock,
-      cover_image: book.cover_image,
-      created_at: book.created_at,   // Thêm trường created_at
-      updated_at: book.updated_at
-    }));
-
+     const formattedBooks = formatBooks(books);
     // Trả về danh sách sách đã được định dạng
     return formattedBooks; // Đảm bảo trả về danh sách đã định dạng
   } catch (error) {
@@ -289,40 +195,51 @@ exports.updateBook = async (id, updatedData) => {
   }
 };
 
-// Xóa một cuốn sách
-exports.deleteBook = async (id) => {
+// // Xóa một cuốn sách
+// exports.deleteBook = async (id) => {
+//   try {
+//       const deletedBook = await bookService.deleteBook(id);
+//       return deletedBook;
+//   } catch (error) {
+//       throw new Error('Error deleting book: ' + error.message);
+//   }
+// };
+// Xóa sách hoặc thay đổi trạng thái sách
+exports.deleteOrHideBook = async (id) => {
   try {
-      const deletedBook = await bookService.deleteBook(id);
-      return deletedBook;
+    const book = await bookService.getBookById(id); // Lấy thông tin sách
+
+    if (!book) {
+      throw new Error('Book not found');
+    }
+
+    // Nếu sách có lượt bán, thay đổi trạng thái thay vì xóa
+    if (book.sales > 0) {
+      const newStatus = book.status === 'visible' ? 'hidden' : 'visible';
+      const updatedBook = await bookService.updateBookStatus(id, newStatus);
+      return {
+        message: `Book status updated to ${newStatus}`,
+        book: updatedBook
+      };
+    }
+
+    // Nếu không có lượt bán, xóa sách
+    const deletedBook = await bookService.deleteBook(id);
+    return {
+      message: 'Book deleted successfully',
+      book: deletedBook
+    };
   } catch (error) {
-      throw new Error('Error deleting book: ' + error.message);
+    throw new Error('Error handling book deletion: ' + error.message);
   }
 };
+
 // tìm kiếm sách theo tiêu đề, tác giả, mô tả
 exports.searchBooks = async (query) => {
   try {
       const books = await bookService.searchBooks(query);
       // Định dạng dữ liệu trước khi trả về
-    const formattedBooks = books.map((book) => ({
-      id: book._id, // ObjectId của sách
-      title: book.title,
-      author: book.author,
-      category: book.categoryID ? {
-          id: book.categoryID._id, // ID danh mục
-          name: book.categoryID.name // Tên danh mục
-      } : null, // Nếu không có categoryID, trả về null
-      publisher: book.publisherID ? {
-          id: book.publisherID._id, // ID nhà xuất bản
-          name: book.publisherID.name // Tên nhà xuất bản
-      } : null, // Nếu không có publisherID, trả về null
-      description: book.description,
-      price: book.price,
-      stock: book.stock,
-      cover_image: book.cover_image,
-      created_at: book.created_at,   // Thêm trường created_at
-      updated_at: book.updated_at
-    }));
-
+    const formattedBooks = formatBooks(books);
     // Trả về danh sách sách đã được định dạng
     return formattedBooks; // Đảm bảo trả về danh sách đã định dạng
   } catch (error) {
@@ -337,26 +254,7 @@ exports.getHotProducts = async (req, res) => {
     const products = await bookService.getHotProducts();
 
     // Định dạng dữ liệu sản phẩm trước khi trả về
-    const formattedProducts = products.map((product) => ({
-      id: product._id, // ObjectId của sản phẩm
-      title: product.title,
-      author: product.author,
-      category: product.categoryID ? {
-          id: product.categoryID._id, // ID danh mục
-          name: product.categoryID.name // Tên danh mục
-      } : null, // Nếu không có categoryID, trả về null
-      publisher: product.publisherID ? {
-          id: product.publisherID._id, // ID nhà xuất bản
-          name: product.publisherID.name // Tên nhà xuất bản
-      } : null, // Nếu không có publisherID, trả về null
-      description: product.description,
-      price: product.price,
-      stock: product.stock,
-      cover_image: product.cover_image,
-      created_at: product.created_at,   // Thêm trường created_at
-      updated_at: product.updated_at
-    }));
-
+    const formattedProducts = formatProducts(products);
     // Trả về danh sách sản phẩm đã được định dạng
     res.status(200).json(formattedProducts);
   } catch (error) {
@@ -371,26 +269,7 @@ exports.getFeaturedProducts = async (req, res) => {
     const products = await bookService.getFeaturedProducts();
 
      // Định dạng dữ liệu sản phẩm trước khi trả về
-     const formattedProducts = products.map((product) => ({
-      id: product._id, // ObjectId của sản phẩm
-      title: product.title,
-      author: product.author,
-      category: product.categoryID ? {
-          id: product.categoryID._id, // ID danh mục
-          name: product.categoryID.name // Tên danh mục
-      } : null, // Nếu không có categoryID, trả về null
-      publisher: product.publisherID ? {
-          id: product.publisherID._id, // ID nhà xuất bản
-          name: product.publisherID.name // Tên nhà xuất bản
-      } : null, // Nếu không có publisherID, trả về null
-      description: product.description,
-      price: product.price,
-      stock: product.stock,
-      cover_image: product.cover_image,
-      created_at: product.created_at,   // Thêm trường created_at
-      updated_at: product.updated_at
-    }));
-
+     const formattedProducts = formatProducts(products);
     // Trả về danh sách sản phẩm đã được định dạng
     res.status(200).json(formattedProducts);
   } catch (error) {
@@ -404,29 +283,21 @@ exports.getLatestBooks = async (req, res) => {
     const products = await bookService.getLatestBooks();
 
     // Định dạng dữ liệu sản phẩm trước khi trả về
-    const formattedProducts = products.map((product) => ({
-      id: product._id, // ObjectId của sản phẩm
-      title: product.title,
-      author: product.author,
-      category: product.categoryID ? {
-          id: product.categoryID._id, // ID danh mục
-          name: product.categoryID.name // Tên danh mục
-      } : null, // Nếu không có categoryID, trả về null
-      publisher: product.publisherID ? {
-          id: product.publisherID._id, // ID nhà xuất bản
-          name: product.publisherID.name // Tên nhà xuất bản
-      } : null, // Nếu không có publisherID, trả về null
-      description: product.description,
-      price: product.price,
-      stock: product.stock,
-      cover_image: product.cover_image,
-      created_at: product.created_at,   // Thêm trường created_at
-      updated_at: product.updated_at
-    }));
-
+    const formattedProducts = formatProducts(products);
     // Trả về danh sách sản phẩm đã được định dạng
     res.status(200).json(formattedProducts);
   } catch (error) {
     res.status(500).json({ error: 'Error fetching latest books: ' + error.message });
+  }
+};
+exports.getLowStockBooks = async () => {
+  try {
+    const books = await bookService.getLowStockBooks();
+    console.log(books)
+    // Định dạng lại dữ liệu trước khi trả về
+    const formattedBooks = formatBooks(books);
+    return formattedBooks;
+  } catch (error) {
+    throw new Error('Error fetching low stock books: ' + error.message);
   }
 };
